quizapp/requirements.txt
Django==4.2.7
django-crispy-forms==2.1
Pillow==10.1.0
django-bootstrap4